% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatCor.R
\name{MatCor}
\alias{MatCor}
\title{Calculate correlation between two matrices}
\usage{
MatCor(X, Y, method, adjust.method = "BH", ...)
}
\arguments{
\item{X}{numaric matrix of data values.}

\item{Y}{numeric matrix of data values. Must have same number of columns as X.}

\item{method}{character value indicating the method to use. Must be one of
"pearson", "spearman", "kendall", or "lm".}

\item{adjust.method}{character value indicating the method to use for
p value adjustement. See \code{\link{p.adjust}}.}

\item{...}{other arguments. If either of "person", "kendal", or "spearman"
is specified, any other arguments in the \code{\link{cor.test}} function can
be used. If "lm" is specified, a design matrix \strong{must} be passed. The
design matrix can be built using the \code{\link{model.matrix}} function.}
}
\value{
A list of data frames. The length of the list equals to the number
of rows of X, and the number of rows of each data fram equals to the number
of rows of Y. Each data frame has 4 columns.

\item{\strong{stat}}{the value of the test statistic.}
\item{\strong{estimate}}{the estimated measure of association. Corresponds to the "r"
for pearson, "rho" for spearman, "tau" for kendal, and coefficient for lm}
\item{\strong{pval}}{the p-value of the test.}
\item{\strong{padj}}{the p-value of the test after multiple test adjustment.}
}
\description{
This funciton calculate correlation between 2 matrices. The 2
matrices must have samples (observations) on columns, and variables to
compute correltaion on rows. The 2 matrices must have same number of
columns. Methods supported are pearson, spearman, kendall, and linear model
with more complex design.
}
\author{
Chenghao Zhu
}
